Rep-Counter/requirements.txt
absl-py==2.3.1
altgraph
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.118.0
flatbuffers==25.9.23
future
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httplib2==0.31.0
idna==3.10
importlib_metadata==8.7.0
keras==3.10.0
libclang==18.1.1
macholib
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
motor==3.7.1
namex==0.1.0
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.2
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
rsa==4.9.1
scipy==1.13.1
six
sniffio==1.3.1
starlette==0.48.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
